Refactoring: Move process termination utility functions to subprocess_util
